Rewrite the option parsing to use getopt_long rather than Xrm.
With this change, option parsing is handled through getopt_long rather
than as X resources. This is an incompatible change from the original
xclip as the new long options start with two dashes. It also changes
the way selection are choosen as now they are checked to have just one
character in the string — at least for now.

With this change, Xresources file support is lost. I'll be deciding
whether to use confuse and switch out of Xresource files for
configuration or if I will write an Xresource parser instead.

Because it's all related, the reading of files has been rewritten so
that it doesn't need to copy the files' names.